requirements.txt
Django==4.2
python-docx==0.8.11
python-pptx==0.6.21
PyMuPDF==1.20.0
pdf2docx==0.5.6
